# exercises/instructor-solutions/hotel/api/openapi.json
{"openapi":"3.1.0","info":{"title":"hotel REST API","license":{"name":"MIT","url":"https://opensource.org/license/mit"},"version":"1.0.0"},"servers":[{"url":"/","description":"Aktueller Server"}],"security":[],"tags":[{"name":"RoomType","description":"HTTP-Ressource für Zimmertypen"}],"paths":{"/api/room-types/{id}":{"get":{"tags":["RoomType"],"summary":"RoomType nach ID lesen","operationId":"findById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64","exclusiveMinimum":0}}],"responses":{"200":{"description":"Ressource gefunden","headers":{"X-Correlation-ID":{"description":"ID zur Zuordnung von Client-Antwort und Server-Log","schema":{"type":"string"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoomTypeResponse"}}}},"404":{"description":"ID ist unbekannt","headers":{"X-Correlation-ID":{"description":"ID zur Zuordnung von Client-Antwort und Server-Log","schema":{"type":"string"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"400":{"description":"ID ist keine positive Zahl","headers":{"X-Correlation-ID":{"description":"ID zur Zuordnung von Client-Antwort und Server-Log","schema":{"type":"string"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"406":{"description":"Angeforderte Repräsentation ist nicht verfügbar","headers":{"X-Correlation-ID":{"description":"ID zur Zuordnung von Client-Antwort und Server-Log","schema":{"type":"string"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Unerwarteter Serverfehler","headers":{"X-Correlation-ID":{"description":"ID zur Zuordnung von Client-Antwort und Server-Log","schema":{"type":"string"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}}},"security":[]},"put":{"tags":["RoomType"],"summary":"RoomType vollständig ersetzen","operationId":"replace","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64","exclusiveMinimum":0}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRoomTypeRequest"}}},"required":true},"responses":{"200":{"description":"Ressource ersetzt","headers":{"X-Correlation-ID":{"description":"ID zur Zuordnung von Client-Antwort und Server-Log","schema":{"type":"string"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoomTypeResponse"}}}},"400":{"description":"Eingabe oder JSON ist ungültig","headers":{"X-Correlation-ID":{"description":"ID zur Zuordnung von Client-Antwort und Server-Log","schema":{"type":"string"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"ID ist unbekannt","headers":{"X-Correlation-ID":{"description":"ID zur Zuordnung von Client-Antwort und Server-Log","schema":{"type":"string"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"409":{"description":"Fachschlüssel oder Datenbankregel kollidiert","headers":{"X-Correlation-ID":{"description":"ID zur Zuordnung von Client-Antwort und Server-Log","schema":{"type":"string"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"415":{"description":"Content-Type ist nicht application/json","headers":{"X-Correlation-ID":{"description":"ID zur Zuordnung von Client-Antwort und Server-Log","schema":{"type":"string"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"406":{"description":"Angeforderte Repräsentation ist nicht verfügbar","headers":{"X-Correlation-ID":{"description":"ID zur Zuordnung von Client-Antwort und Server-Log","schema":{"type":"string"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Unerwarteter Serverfehler","headers":{"X-Correlation-ID":{"description":"ID zur Zuordnung von Client-Antwort und Server-Log","schema":{"type":"string"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}}},"security":[]},"delete":{"tags":["RoomType"],"summary":"RoomType löschen","description":"Idempotent: Eine unbekannte ID liefert ebenfalls 204.","operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64","exclusiveMinimum":0}}],"responses":{"204":{"description":"Ressource ist gelöscht","headers":{"X-Correlation-ID":{"description":"ID zur Zuordnung von Client-Antwort und Server-Log","schema":{"type":"string"}}}},"409":{"description":"Referenzierende Daten verhindern das Löschen","headers":{"X-Correlation-ID":{"description":"ID zur Zuordnung von Client-Antwort und Server-Log","schema":{"type":"string"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"400":{"description":"ID ist keine positive Zahl","headers":{"X-Correlation-ID":{"description":"ID zur Zuordnung von Client-Antwort und Server-Log","schema":{"type":"string"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"406":{"description":"Angeforderte Repräsentation ist nicht verfügbar","headers":{"X-Correlation-ID":{"description":"ID zur Zuordnung von Client-Antwort und Server-Log","schema":{"type":"string"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Unerwarteter Serverfehler","headers":{"X-Correlation-ID":{"description":"ID zur Zuordnung von Client-Antwort und Server-Log","schema":{"type":"string"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}}},"security":[]}},"/api/room-types":{"get":{"tags":["RoomType"],"summary":"Zimmertypen auflisten","operationId":"findAll","responses":{"200":{"description":"Nach ID sortierte Liste","headers":{"X-Correlation-ID":{"description":"ID zur Zuordnung von Client-Antwort und Server-Log","schema":{"type":"string"}}},"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RoomTypeResponse"}}}}},"406":{"description":"Angeforderte Repräsentation ist nicht verfügbar","headers":{"X-Correlation-ID":{"description":"ID zur Zuordnung von Client-Antwort und Server-Log","schema":{"type":"string"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Unerwarteter Serverfehler","headers":{"X-Correlation-ID":{"description":"ID zur Zuordnung von Client-Antwort und Server-Log","schema":{"type":"string"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}}},"security":[]},"post":{"tags":["RoomType"],"summary":"RoomType anlegen","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRoomTypeRequest"}}},"required":true},"responses":{"201":{"description":"Ressource angelegt","headers":{"X-Correlation-ID":{"description":"ID zur Zuordnung von Client-Antwort und Server-Log","schema":{"type":"string"}},"Location":{"description":"URL der neuen Ressource","style":"simple","schema":{"type":"string","format":"uri"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoomTypeResponse"}}}},"400":{"description":"Eingabe oder JSON ist ungültig","headers":{"X-Correlation-ID":{"description":"ID zur Zuordnung von Client-Antwort und Server-Log","schema":{"type":"string"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"409":{"description":"type_code ist vergeben oder eine Datenbankregel wird verletzt","headers":{"X-Correlation-ID":{"description":"ID zur Zuordnung von Client-Antwort und Server-Log","schema":{"type":"string"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"415":{"description":"Content-Type ist nicht application/json","headers":{"X-Correlation-ID":{"description":"ID zur Zuordnung von Client-Antwort und Server-Log","schema":{"type":"string"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"406":{"description":"Angeforderte Repräsentation ist nicht verfügbar","headers":{"X-Correlation-ID":{"description":"ID zur Zuordnung von Client-Antwort und Server-Log","schema":{"type":"string"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Unerwarteter Serverfehler","headers":{"X-Correlation-ID":{"description":"ID zur Zuordnung von Client-Antwort und Server-Log","schema":{"type":"string"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiError"}}}}},"security":[]}},"/api/health":{"get":{"tags":["health-controller"],"summary":"Anwendungsstatus lesen","operationId":"health","responses":{"200":{"description":"Anwendung ist erreichbar","headers":{"X-Correlation-ID":{"description":"ID zur Zuordnung von Client-Antwort und Server-Log","schema":{"type":"string"}}},"content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"406":{"description":"Angeforderte Repräsentation ist nicht verfügbar","headers":{"X-Correlation-ID":{"description":"ID zur Zuordnung von Client-Antwort und Server-Log","schema":{"type":"string"}}},"content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}}},"security":[]}}},"components":{"schemas":{"UpdateRoomTypeRequest":{"type":"object","description":"Eingabe für Zimmertypen","properties":{"typeCode":{"type":"string","description":"Eindeutiger Typcode","example":"NX","maxLength":10,"minLength":1},"name":{"type":"string","description":"Name des Zimmertyps","example":"Testzimmer","maxLength":80,"minLength":1},"capacity":{"type":"integer","format":"int32","description":"Maximale Belegung","example":2,"maximum":6,"minimum":1},"standardPriceCents":{"type":"integer","format":"int32","description":"Standardpreis in Cent","example":14900,"minimum":0}},"required":["capacity","name","standardPriceCents","typeCode"]},"RoomTypeResponse":{"type":"object","description":"Öffentliche Darstellung von Zimmertypen","properties":{"id":{"type":"integer","format":"int64","description":"Vom Server vergebene ID","example":1,"readOnly":true},"typeCode":{"type":"string","description":"Eindeutiger Typcode","example":"NX"},"name":{"type":"string","description":"Name des Zimmertyps","example":"Testzimmer"},"capacity":{"type":"integer","format":"int32","description":"Maximale Belegung","example":2},"standardPriceCents":{"type":"integer","format":"int32","description":"Standardpreis in Cent","example":14900}},"required":["capacity","id","name","standardPriceCents","typeCode"]},"ApiError":{"type":"object","description":"Stabiler Fehlerkörper für alle API-Fehler","properties":{"code":{"type":"string","example":"VALIDATION_FAILED"},"message":{"type":"string","example":"Eingabe ist ungültig"},"correlationId":{"type":"string","example":"request-42"},"fields":{"type":"object","additionalProperties":{"type":"string"},"description":"Fehler je Eingabefeld, sonst leer"}},"required":["code","correlationId","fields","message"]},"CreateRoomTypeRequest":{"type":"object","description":"Eingabe für Zimmertypen","properties":{"typeCode":{"type":"string","description":"Eindeutiger Typcode","example":"NX","maxLength":10,"minLength":1},"name":{"type":"string","description":"Name des Zimmertyps","example":"Testzimmer","maxLength":80,"minLength":1},"capacity":{"type":"integer","format":"int32","description":"Maximale Belegung","example":2,"maximum":6,"minimum":1},"standardPriceCents":{"type":"integer","format":"int32","description":"Standardpreis in Cent","example":14900,"minimum":0}},"required":["capacity","name","standardPriceCents","typeCode"]}}}}